fix: user can only be deleted when there is no registered

+++ src/services/userService.ts
@@ -2,13 +2,16 @@ import { IUserServiceInterface } from "./interfaces/userServiceInterface";
 import bcrypt from "bcryptjs";
 import UserDTO, { SecureUser } from "../models/user";
 import { IUserRepository } from "../repositories/interfaces/userRepositoryInterface";
+import { IAnnouncementRepository } from "../repositories/interfaces/announcementRepositoryInterface";
 
 
 export class UserService implements IUserServiceInterface {
     private userRepository: IUserRepository;
+    private announcementRepository: IAnnouncementRepository;
 
-    constructor(iUserRepository: IUserRepository) {
+    constructor(iUserRepository: IUserRepository, iAnnouncementRepository: IAnnouncementRepository) {
         this.userRepository = iUserRepository;
+        this.announcementRepository = iAnnouncementRepository;
     }
 
     async getById( id: number ): Promise<SecureUser> {
@@ -55,6 +58,10 @@ export class UserService implements IUserServiceInterface {
         const userExists = await this.userRepository.selectOne({ id })
 
         if (!userExists) throw new Error("Usuario não encontrado!");
+
+        const advertiserId = id;
+        const announcementExist = await this.announcementRepository.selectOne({ advertiserId })
+        if(announcementExist) throw new Error ("Usuário não pode ser deletado, pois possui anúncio(s) cadastrado(s).");
 
         const msg = await this.userRepository.delete( id )
